test: move tests for `build_serialized_output`
